feat(proposal): getLogs from snapshot to min(startBlock, currentBlock)

+++ plugins/maciVoting/artifacts/MaciVoting.sol.tsx
@@ -257,8 +257,8 @@ export const MaciVotingAbi = [
               },
               {
                 name: "snapshotBlock",
-                type: "uint64",
-                internalType: "uint64",
+                type: "uint256",
+                internalType: "uint256",
               },
               {
                 name: "minVotingPower",
